+++ demo/history/main.c
@@ -133,9 +133,9 @@ int main(int argc, char **argv) {
 		return 1;
 	}
 	client->logger_in =
-		rpclogger_new(rpclogger_func_stderr, "<= ", logsiz, conf.verbose);
+		rpclogger_new(&rpclogger_stderr_funcs, "<= ", logsiz, conf.verbose);
 	client->logger_out =
-		rpclogger_new(rpclogger_func_stderr, "=> ", logsiz, conf.verbose);
+		rpclogger_new(&rpclogger_stderr_funcs, "=> ", logsiz, conf.verbose);
 
 	struct rpchandler_history_records *records[2] = {
 		&(struct rpchandler_history_records){.name = "records_log",